+++ Sources/FinnhubSwift/Models/PriceTarget.swift
@@ -3,8 +3,8 @@ import Foundation
 public struct PriceTarget: Mappable {
     public var lastUpdated: String
     public var symbol: String
-    public var targetHigh: Double
-    public var targetLow: Double
-    public var targetMean: Double
-    public var targetMedian: Double
+    public var targetHigh: Float
+    public var targetLow: Float
+    public var targetMean: Float
+    public var targetMedian: Float
 }
